Add files via upload
A classification problem on UCI adult data set. 
Evaluate the Decision Tree model using cross-validation with the default parameters. 
Tune the parameters of the models using GridSearchCV for max_depth with values range(2, 20, 2). What is the best parameter and best validation score from the grid search? 
 Visualize the dependence of the validation score on the parameter max_depth. 
 Tune the parameters of the models using GridSearchCV for ccp_alpha with values np.linspace(0., 0.03, 10). What is the best parameter and best validation score from the grid search? 
Visualize the dependence of the validation score on the parameter ccp_alpha. (5 points)
Compare the best cross-validation scores of the two grid searches. Which model is better? What is the performance of the best tree model on the testing set? 
 Plot the best model’s tree structure. 
 Use the same train test split. Use proper preprocessing. Using the training set to train a Gradient Boosting model with the default parameters. No needs to cross-validate or grid search. What is the model’s performance on the testing set? (10 points)
Compare the top 10 important features of decision tree and gradient boosting. How much do they agree?